Update bootstrap-remote-tabs.js
Changed location of tabContainer.html(data) so that content exists prior to callbackFn for better binding.

Added in tabParent as the parent of all tab panels to enable a an empty request on change (lines 130-132). Both the tabParent value needs to exist and on each container the class remove-rd so that the remove action is intentional, especially given this is undesirable in the case where the tab doesn't need to be refreshed.

The point of that was to solve the issue of binding events to hidden tab content.